Incorrect messages/Minor errors
Incorrect messages stating that a title had to be added to newly added events but there were already titles, incorrect message stating that the add event boxes hadn't been completely filled when they had been. New events that were added being duplicated, system messages saying adding a event were also duplicated. When different boxes were clicked on the calendar, it took the user to a part with the daily schedule and had a pre-added date however showed the current date not the date corresponding to the box clicked on. All of this is fixed